Reuse plots.test-support.ts instead of duplicating helpers (warren-8e04)
Follow-up: the split files now import the shared seam stubs from the
existing sibling helper plots.test-support.ts (adding the two missing
helpers, fakeCreator + summary, there) rather than introducing a
parallel plots.test-helpers.ts. Avoids the copy-paste jscpd flagged.
