+++ src/server/handlers/plots.needs-attention.test.ts
@@ -5,7 +5,7 @@ import type { PlotAggregator, PlotSummary } from "../../plots/index.ts";
 import { NO_AUTH } from "../auth.ts";
 import { startServer } from "../server.ts";
 import type { ServeHandle } from "../types.ts";
-import { depsFor, silentLogger, summary, tcpUrl } from "./plots.test-helpers.ts";
+import { depsFor, silentLogger, summary, tcpUrl } from "./plots.test-support.ts";
 
 describe("GET /plots?filter=needs_attention", () => {
 	let db: WarrenDb;
